[STAD-938] Extract the Grid FS upload path into its own operation
Writing an upload to Grid FS spans three steps which only make sense
together: draining the request into a temporary file, comparing what
arrived against the announced content range, and handing the completed
file to storage. They sat among the lookups and cleanups of
GridFsStorageService, which left that class covering two unrelated
concerns.

Move them into a GridFsUploadOperation, mirroring the GridFsCleanupOperation
which already lives beside it. The temporary file path is now computed once
by the caller and passed in, instead of being derived again in two places.

No behaviour change; GridFSStorageIT covers the moved path.
